+++ frontend/src/service/@shared/hooks/queries/review/index.ts
@@ -1,6 +1,4 @@
-export { default as useGetReviewForm } from './useGetReviewForm';
-export { default as useGetReviews } from './useGetReviews';
-
-export { default as useCreateReviewForm } from './useCreateReviewForm';
-export { default as useCreateReviewAnswer } from './useCreateReviewAnswer';
-export { default as useUpdateReviewForm } from './useUpdateReviewForm';
+export * from './useGet';
+export * from './useCreate';
+export * from './useUpdate';
+export * from './useDelete';
