+++ frontend/src/service/@shared/hooks/queries/review/useCreate.ts
@@ -5,16 +5,26 @@ import { UpdateReviewFormResponse, UseCustomMutationOptions } from 'service/@sha
 import reviewAPI from 'service/@shared/api/review';
 import { QUERY_KEY } from 'service/@shared/constants';
 
-function useUpdateReviewForm(mutationOptions?: UseCustomMutationOptions<UpdateReviewFormResponse>) {
+function useCreateReviewForm(mutationOptions?: UseCustomMutationOptions<UpdateReviewFormResponse>) {
   const queryClient = useQueryClient();
 
-  return useMutation(reviewAPI.updateForm, {
+  return useMutation(reviewAPI.createForm, {
     onSuccess: () => {
       queryClient.invalidateQueries([QUERY_KEY.DATA.REVIEW_FORM]);
+    },
+    ...mutationOptions,
+  });
+}
+
+function useCreateReviewAnswer(mutationOptions?: UseCustomMutationOptions<null>) {
+  const queryClient = useQueryClient();
+
+  return useMutation(reviewAPI.submitAnswer, {
+    onSuccess: () => {
       queryClient.invalidateQueries([QUERY_KEY.DATA.REVIEW]);
     },
     ...mutationOptions,
   });
 }
 
-export default useUpdateReviewForm;
+export { useCreateReviewForm, useCreateReviewAnswer };
