+++ frontend/src/service/@shared/hooks/queries/review/useGet.ts
@@ -1,0 +1,58 @@
+import { useQuery, UseQueryOptions } from 'react-query';
+
+import { ErrorResponse } from 'service/@shared/types';
+import {
+  GetReviewFormResponse,
+  GetReviewResponse,
+  GetReviewsResponse,
+} from 'service/@shared/types/review';
+
+import reviewAPI from 'service/@shared/api/review';
+import { QUERY_KEY } from 'service/@shared/constants';
+
+function useGetReviewForm(
+  reviewFormCode: string,
+  queryOptions?: UseQueryOptions<GetReviewFormResponse, ErrorResponse>,
+) {
+  return useQuery<GetReviewFormResponse, ErrorResponse>(
+    [QUERY_KEY.DATA.REVIEW_FORM, QUERY_KEY.API.GET_REVIEW_FORM, { reviewFormCode }],
+    () => reviewAPI.getForm(reviewFormCode),
+    {
+      suspense: true,
+      useErrorBoundary: false,
+      ...queryOptions,
+    },
+  );
+}
+
+function useGetReviews(
+  reviewFormCode: string,
+  queryOptions?: UseQueryOptions<GetReviewsResponse, ErrorResponse>,
+) {
+  return useQuery<GetReviewsResponse, ErrorResponse>(
+    [QUERY_KEY.DATA.REVIEW, QUERY_KEY.API.GET_REVIEWS, { reviewFormCode }],
+    () => reviewAPI.getReviews(reviewFormCode),
+    {
+      suspense: true,
+      useErrorBoundary: false,
+      ...queryOptions,
+    },
+  );
+}
+
+function useGetReview(
+  reviewId: number,
+  queryOptions?: UseQueryOptions<GetReviewResponse, ErrorResponse>,
+) {
+  return useQuery<GetReviewResponse, ErrorResponse>(
+    [QUERY_KEY.DATA.REVIEW, QUERY_KEY.API.GET_REVIEW, { reviewId }],
+    () => reviewAPI.getReview(reviewId),
+    {
+      suspense: true,
+      useErrorBoundary: false,
+      ...queryOptions,
+    },
+  );
+}
+
+export { useGetReviewForm, useGetReviews, useGetReview };
