+++ frontend/src/service/community/pages/MyPage/useMyPageQueries.ts
@@ -1,7 +1,6 @@
 import useGetMyReviewForms from 'service/@shared/hooks/queries/profile/useGetMyReviewForms';
 import useGetMyReviews from 'service/@shared/hooks/queries/profile/useGetMyReviews';
-import useDeleteReview from 'service/@shared/hooks/queries/review/useDeleteReview';
-import useDeleteReviewForm from 'service/@shared/hooks/queries/review/useDeleteReviewForm';
+import { useDeleteReview, useDeleteReviewForm } from 'service/@shared/hooks/queries/review';
 import useGetUserProfile from 'service/@shared/hooks/queries/user/useGetUserProfile';
 
 function useMyPageQueries() {
